refactor(challenges): extract tournament panel and challenge config state
