+++ src/features/challenges/components/challenge-tournament-panel.tsx
@@ -1,5 +1,5 @@
 import Feather from '@expo/vector-icons/Feather';
-import { useEffect, useMemo, useState, type ReactNode } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { Alert, Pressable, TextInput, View } from 'react-native';
 import { Button, Card, Spinner } from 'heroui-native';
 
@@ -10,14 +10,15 @@ import type {
   ChallengeTeam,
   TournamentMatch,
   TournamentMatchInput,
-  TournamentMatchStatus,
 } from '../types/challenge.model';
 import {
   useTournamentAdminActions,
   useTournamentMatches,
   useTournamentScores,
 } from '../hooks/use-configure-challenges';
-import { formatChallengeDate } from '../utils/challenge-config-utils';
+import { MatchCard } from './tournament-match-card';
+import { MatchForm, emptyMatchForm, type TournamentMatchForm } from './tournament-match-form';
+import { buildStandings, groupMatches } from '../utils/tournament-utils';
 
 type TournamentTab = 'fixtures' | 'standings' | 'scores';
 
@@ -31,13 +32,6 @@ const inputStyle = {
   fontSize: 15,
   color: mflColors.text,
 };
-
-const STATUS_OPTIONS: Array<{ value: TournamentMatchStatus; label: string }> = [
-  { value: 'scheduled', label: 'Scheduled' },
-  { value: 'live', label: 'Live' },
-  { value: 'completed', label: 'Completed' },
-  { value: 'cancelled', label: 'Cancelled' },
-];
 
 interface ChallengeTournamentPanelProps {
   leagueId: string;
@@ -336,338 +330,5 @@ export function ChallengeTournamentPanel({
         </View>
       ) : null}
     </Card>
-  );
-}
-
-interface TournamentMatchForm {
-  roundNumber: string;
-  roundName: string;
-  team1Id: string;
-  team2Id: string;
-  startDate: string;
-  status: TournamentMatchStatus;
-  score1: string;
-  score2: string;
-}
-
-function emptyMatchForm(): TournamentMatchForm {
-  return {
-    roundNumber: '1',
-    roundName: 'Group Stage',
-    team1Id: '',
-    team2Id: '',
-    startDate: new Date().toISOString().slice(0, 10),
-    status: 'scheduled',
-    score1: '0',
-    score2: '0',
-  };
-}
-
-function MatchForm({
-  form,
-  teams,
-  isSaving,
-  editing,
-  onChange,
-  onCancel,
-  onSubmit,
-}: {
-  form: TournamentMatchForm;
-  teams: ChallengeTeam[];
-  isSaving: boolean;
-  editing: boolean;
-  onChange: (patch: Partial<TournamentMatchForm>) => void;
-  onCancel: () => void;
-  onSubmit: () => void;
-}) {
-  return (
-    <View className="rounded-xl border border-default-200 p-3 gap-3">
-      <AppText className="text-base font-bold text-foreground">
-        {editing ? 'Edit Match' : 'Add Match'}
-      </AppText>
-
-      <View className="flex-row gap-3">
-        <Field label="Round" className="flex-1">
-          <TextInput
-            style={inputStyle}
-            value={form.roundNumber}
-            onChangeText={(value) => onChange({ roundNumber: value })}
-            keyboardType="numeric"
-            placeholder="1"
-            placeholderTextColor={mflColors.textMuted}
-          />
-        </Field>
-        <Field label="Date" className="flex-1">
-          <TextInput
-            style={inputStyle}
-            value={form.startDate}
-            onChangeText={(value) => onChange({ startDate: value })}
-            placeholder="YYYY-MM-DD"
-            placeholderTextColor={mflColors.textMuted}
-            keyboardType="numbers-and-punctuation"
-            maxLength={10}
-          />
-        </Field>
-      </View>
-
-      <Field label="Round Name">
-        <TextInput
-          style={inputStyle}
-          value={form.roundName}
-          onChangeText={(value) => onChange({ roundName: value })}
-          placeholder="Group Stage"
-          placeholderTextColor={mflColors.textMuted}
-        />
-      </Field>
-
-      <TeamPicker
-        label="Team 1"
-        teams={teams}
-        value={form.team1Id}
-        disabledTeamId={form.team2Id}
-        onChange={(team1Id) => onChange({ team1Id })}
-      />
-      <TeamPicker
-        label="Team 2"
-        teams={teams}
-        value={form.team2Id}
-        disabledTeamId={form.team1Id}
-        onChange={(team2Id) => onChange({ team2Id })}
-      />
-
-      <View className="gap-2">
-        <AppText className="text-xs font-semibold text-muted uppercase">Status</AppText>
-        <View className="flex-row flex-wrap gap-2">
-          {STATUS_OPTIONS.map((status) => (
-            <Pressable
-              key={status.value}
-              onPress={() => onChange({ status: status.value })}
-              className="rounded-full border px-3 py-2"
-              style={{
-                backgroundColor: form.status === status.value ? mflColors.brand : mflColors.card,
-                borderColor: form.status === status.value ? mflColors.brand : mflColors.border,
-              }}
-            >
-              <AppText
-                className="text-xs font-bold"
-                style={{ color: form.status === status.value ? '#fff' : mflColors.text }}
-              >
-                {status.label}
-              </AppText>
-            </Pressable>
-          ))}
-        </View>
-      </View>
-
-      {form.status !== 'scheduled' || editing ? (
-        <View className="flex-row gap-3">
-          <Field label="Score 1" className="flex-1">
-            <TextInput
-              style={inputStyle}
-              value={form.score1}
-              onChangeText={(value) => onChange({ score1: value })}
-              keyboardType="numeric"
-              placeholder="0"
-              placeholderTextColor={mflColors.textMuted}
-            />
-          </Field>
-          <Field label="Score 2" className="flex-1">
-            <TextInput
-              style={inputStyle}
-              value={form.score2}
-              onChangeText={(value) => onChange({ score2: value })}
-              keyboardType="numeric"
-              placeholder="0"
-              placeholderTextColor={mflColors.textMuted}
-            />
-          </Field>
-        </View>
-      ) : null}
-
-      <View className="flex-row gap-3">
-        <Button variant="secondary" size="md" onPress={onCancel} className="flex-1">
-          <Button.Label>Cancel</Button.Label>
-        </Button>
-        <Button variant="primary" size="md" onPress={onSubmit} isDisabled={isSaving} className="flex-1">
-          {isSaving ? <Spinner size="sm" /> : <Button.Label>Save Match</Button.Label>}
-        </Button>
-      </View>
-    </View>
-  );
-}
-
-function Field({
-  label,
-  className,
-  children,
-}: {
-  label: string;
-  className?: string;
-  children: ReactNode;
-}) {
-  return (
-    <View className={`gap-2 ${className ?? ''}`}>
-      <AppText className="text-xs font-semibold text-muted uppercase">{label}</AppText>
-      {children}
-    </View>
-  );
-}
-
-function TeamPicker({
-  label,
-  teams,
-  value,
-  disabledTeamId,
-  onChange,
-}: {
-  label: string;
-  teams: ChallengeTeam[];
-  value: string;
-  disabledTeamId: string;
-  onChange: (teamId: string) => void;
-}) {
-  return (
-    <View className="gap-2">
-      <AppText className="text-xs font-semibold text-muted uppercase">{label}</AppText>
-      <View className="flex-row flex-wrap gap-2">
-        {teams.map((team) => {
-          const selected = value === team.teamId;
-          const disabled = disabledTeamId === team.teamId;
-          return (
-            <Pressable
-              key={team.teamId}
-              disabled={disabled}
-              onPress={() => onChange(team.teamId)}
-              className="rounded-full border px-3 py-2"
-              style={{
-                opacity: disabled ? 0.4 : 1,
-                backgroundColor: selected ? mflColors.brand : mflColors.card,
-                borderColor: selected ? mflColors.brand : mflColors.border,
-              }}
-            >
-              <AppText
-                className="text-xs font-bold"
-                style={{ color: selected ? '#fff' : mflColors.text }}
-              >
-                {team.teamName}
-              </AppText>
-            </Pressable>
-          );
-        })}
-      </View>
-    </View>
-  );
-}
-
-function MatchCard({
-  match,
-  onEdit,
-  onDelete,
-}: {
-  match: TournamentMatch;
-  onEdit: () => void;
-  onDelete: () => void;
-}) {
-  return (
-    <View className="rounded-xl border border-default-200 overflow-hidden">
-      <View className="flex-row items-center justify-between bg-default-100 px-3 py-2">
-        <AppText className="text-xs text-muted">
-          {formatChallengeDate(match.startTime)} - {match.status}
-        </AppText>
-        <View className="flex-row gap-2">
-          <Button variant="secondary" size="sm" onPress={onEdit}>
-            <Button.Label>Edit</Button.Label>
-          </Button>
-          <Button variant="secondary" size="sm" onPress={onDelete}>
-            <Button.Label style={{ color: mflColors.danger }}>Delete</Button.Label>
-          </Button>
-        </View>
-      </View>
-      <View className="flex-row items-center p-4 gap-3">
-        <TeamMatchSide name={match.team1Name || 'TBD'} />
-        <View className="items-center px-2">
-          <AppText className="text-xl font-bold text-foreground">
-            {match.status === 'scheduled' ? 'vs' : `${match.score1} - ${match.score2}`}
-          </AppText>
-        </View>
-        <TeamMatchSide name={match.team2Name || 'TBD'} />
-      </View>
-    </View>
-  );
-}
-
-function TeamMatchSide({ name }: { name: string }) {
-  return (
-    <View className="flex-1 items-center gap-2">
-      <View className="w-10 h-10 rounded-full bg-default-100 items-center justify-center">
-        <Feather name="shield" size={18} color={mflColors.brand} />
-      </View>
-      <AppText className="text-sm font-semibold text-foreground text-center" numberOfLines={2}>
-        {name}
-      </AppText>
-    </View>
-  );
-}
-
-function groupMatches(matches: TournamentMatch[]) {
-  const grouped = new Map<string, TournamentMatch[]>();
-  matches.forEach((match) => {
-    const roundName = match.roundName || `Round ${match.roundNumber}`;
-    grouped.set(roundName, [...(grouped.get(roundName) ?? []), match]);
-  });
-  return Array.from(grouped.entries());
-}
-
-function buildStandings(teams: ChallengeTeam[], matches: TournamentMatch[]) {
-  const stats = new Map(
-    teams.map((team) => [
-      team.teamId,
-      {
-        teamId: team.teamId,
-        teamName: team.teamName,
-        played: 0,
-        won: 0,
-        drawn: 0,
-        lost: 0,
-        points: 0,
-        goalsFor: 0,
-        goalDifference: 0,
-      },
-    ]),
-  );
-
-  matches.forEach((match) => {
-    if (match.status !== 'completed' || !match.team1Id || !match.team2Id) return;
-    const team1 = stats.get(match.team1Id);
-    const team2 = stats.get(match.team2Id);
-    if (!team1 || !team2) return;
-
-    team1.played += 1;
-    team2.played += 1;
-    team1.goalsFor += match.score1;
-    team2.goalsFor += match.score2;
-    team1.goalDifference += match.score1 - match.score2;
-    team2.goalDifference += match.score2 - match.score1;
-
-    if (match.score1 > match.score2) {
-      team1.won += 1;
-      team1.points += 3;
-      team2.lost += 1;
-      team2.points += 1;
-    } else if (match.score2 > match.score1) {
-      team2.won += 1;
-      team2.points += 3;
-      team1.lost += 1;
-      team1.points += 1;
-    } else {
-      team1.drawn += 1;
-      team2.drawn += 1;
-      team1.points += 2;
-      team2.points += 2;
-    }
-  });
-
-  return Array.from(stats.values()).sort(
-    (a, b) => b.points - a.points || b.goalDifference - a.goalDifference || b.goalsFor - a.goalsFor,
   );
 }
